refactor: 목함수 파일 분리 추가
